+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.3.0] - 2020-10-01
 ### Changed
 - Applied the default Syn project meta files ([#70])
 - Rework documentation ([#71])
@@ -46,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial implementation
 
-[Unreleased]: https://github.com/projectsyn/lieutenant-api/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/projectsyn/lieutenant-api/compare/v0.3.0...HEAD
 [v0.1.1]: https://github.com/projectsyn/lieutenant-api/releases/tag/v0.1.1
 [v0.1.2]: https://github.com/projectsyn/lieutenant-api/releases/tag/v0.1.2
 [v0.1.3]: https://github.com/projectsyn/lieutenant-api/releases/tag/v0.1.3
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v0.1.5]: https://github.com/projectsyn/lieutenant-api/releases/tag/v0.1.5
 [v0.1.5]: https://github.com/projectsyn/lieutenant-api/releases/tag/v0.1.5
 [v0.2.0]: https://github.com/projectsyn/lieutenant-api/releases/tag/v0.2.0
+[v0.3.0]: https://github.com/projectsyn/lieutenant-api/releases/tag/v0.3.0
 
 [#55]: https://github.com/projectsyn/lieutenant-api/pull/55
 [#61]: https://github.com/projectsyn/lieutenant-api/pull/61